+++ tests/apps/slack_integration/api/oauth_test.py
@@ -1,7 +1,7 @@
 import pytest
 import responses
 
-from busy_beaver.apps.oauth_integrations.oauth_providers.slack import SlackOAuthFlow
+from busy_beaver.apps.slack_integration.oauth_flow import SlackOAuthFlow
 from busy_beaver.models import SlackInstallation
 from tests._utilities import FakeSlackClient
